Resolve import errors by extracting utility
