+++ src/core/graph/widgets/dynamicWidgets.ts
@@ -206,16 +206,6 @@ const dynamicInputs: Record<
   COMFY_AUTOGROW_V3: applyAutogrow,
   COMFY_MATCHTYPE_V3: applyMatchType
 }
-const dynamicTypeResolvers: Record<
-  string,
-  (inputSpec: InputSpecV2) => string[]
-> = {
-  COMFY_AUTOGROW_V3: resolveAutogrowType,
-  COMFY_MATCHTYPE_V3: (input) =>
-    zMatchTypeOptions
-      .safeParse(input)
-      .data?.template?.allowed_types?.split(',') ?? []
-}
 
 export function applyDynamicInputs(
   node: LGraphNode,
@@ -225,26 +215,6 @@ export function applyDynamicInputs(
   //TODO: move parsing/validation of inputSpec here?
   dynamicInputs[inputSpec.type](node, inputSpec)
   return true
-}
-
-export function resolveInputType(input: InputSpecV2): string[] {
-  return input.type in dynamicTypeResolvers
-    ? dynamicTypeResolvers[input.type](input)
-    : input.type.split(',')
-}
-
-function resolveAutogrowType(rawSpec: InputSpecV2): string[] {
-  const { input } = zAutogrowOptions.safeParse(rawSpec).data?.template ?? {}
-
-  const inputTypes: (Record<string, InputSpec> | undefined)[] = [
-    input?.required,
-    input?.optional
-  ]
-  return inputTypes.flatMap((inputType) =>
-    Object.entries(inputType ?? {}).flatMap(([name, v]) =>
-      resolveInputType(transformInputSpecV1ToV2(v, { name }))
-    )
-  )
 }
 
 function spliceInputs(
